requirements.txt
beniget==0.4.1
Cython==0.29.27
gast==0.5.3
numpy==1.21.5
ply==3.11
pybind11==2.9.1
python-dateutil==2.8.2
pythran==0.11.0
scipy==1.7.3
six==1.16.0
